move to gesture_core
