+++ integrated_tracker/main.py
@@ -1,15 +1,15 @@
 import cv2
 import json
 import mediapipe as mp
-import numpy as np
 import time
 import platform
 from collections import deque
 
 from gesture_core import (
     instructions,
-    classify_gesture,
-    execute_gesture_action
+    update_trajectory,
+    draw_trajectory_on_frame,
+    process_hand_gesture
 )
 from my_gui import show_guidline
 from recognizer import SwipeRecognizer
@@ -45,93 +45,6 @@ def check_OS():
     """
     os_name = platform.system()
     return 1 if os_name == "Windows" else 0 if os_name == "Darwin" else -1
-
-def update_trajectory(trajectory, hand_landmarks, frame_shape):
-    """
-    Calculate the center of the hand based on landmarks and append it to the trajectory.
-
-    Args:
-        trajectory (deque): A deque storing recent hand center positions.
-        hand_landmarks (List[Landmark]): List of 21 hand landmark points.
-        frame_shape (tuple): Shape of the frame (height, width, channels).
-
-    Returns:
-        tuple: The calculated center (cx, cy) of the hand.
-
-    손 랜드마크를 기반으로 손 중심 좌표를 계산하여 trajectory에 저장합니다.
-
-    반환:
-        tuple: 손 중심 좌표 (cx, cy)
-    """
-    h, w, _ = frame_shape
-    cx = int(np.mean([lm.x * w for lm in hand_landmarks]))
-    cy = int(np.mean([lm.y * h for lm in hand_landmarks]))
-    trajectory.append((cx, cy))
-    return cx, cy
-
-
-def draw_trajectory_on_frame(frame, trajectory):
-    """
-    Draw the trajectory of hand movement on the frame.
-
-    Args:
-        frame (np.ndarray): The current video frame.
-        trajectory (deque): A deque containing past hand positions.
-
-    손 이동 궤적을 프레임 위에 선으로 시각화합니다.
-    """
-    for i in range(1, len(trajectory)):
-        cv2.line(frame, trajectory[i - 1], trajectory[i], (255, 0, 0), 2)
-
-
-def should_update_gesture(gesture_timestamp, hold_duration):
-    """
-    Check if the cooldown period for gesture recognition has passed.
-
-    Args:
-        gesture_timestamp (float): Timestamp of the last recognized gesture.
-        hold_duration (float): Cooldown duration in seconds.
-
-    Returns:
-        bool: True if gesture can be updated, False otherwise.
-
-    이전 제스처 인식 이후 쿨다운 시간이 지났는지를 확인합니다.
-
-    반환:
-        bool: 제스처 갱신 가능 여부
-    """
-    cooldown_remaining = hold_duration - (time.time() - gesture_timestamp)
-    return cooldown_remaining <= 0
-
-
-def process_hand_gesture(hand_landmarks, trajectory, gesture_timestamp, hold_duration, os_name):
-    """
-    Determine the gesture from current hand data and handle cooldown logic.
-
-    Args:
-        hand_landmarks (List[Landmark]): List of current hand landmarks.
-        trajectory (deque): Stored hand movement trajectory.
-        gesture_timestamp (float): Timestamp of last gesture recognition.
-        hold_duration (float): Cooldown duration in seconds.
-        os_name (int): OS identifier (1: Windows, 0: macOS, -1: others)
-
-    Returns:
-        tuple: (recognized gesture or None, updated gesture_timestamp)
-
-    현재 손 상태로부터 제스처를 판단하고, 쿨다운 조건을 충족하면 제스처를 실행합니다.
-
-    반환:
-        tuple: (인식된 제스처 or None, 갱신된 gesture_timestamp)
-    """
-    gesture_candidate = classify_gesture(hand_landmarks, trajectory)
-
-    if should_update_gesture(gesture_timestamp, hold_duration):
-        if gesture_candidate:
-            gesture_timestamp = time.time()
-            execute_gesture_action(gesture_candidate, os_name)
-            return gesture_candidate, gesture_timestamp
-
-    return None, gesture_timestamp
 
 
 def send_gesture_via_socket(gesture, is_swipe, last_sent, last_time):
